requirements.txt
google-genai
playwright
jinja2
requests
